@@ -1,16 +1,20 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
+import React from "react";
+import ReactDOM from "react-dom/client";
+import { BrowserRouter } from "react-router-dom";
 
-import './index.css';
-import App from './App';
-import ProductsProvider from './context/products-context';
+import "./index.css";
+import App from "./App";
+// import ProductsProvider from './context/products-context';
+import configureProductsStore from "./hooks-store/products-store";
 
-ReactDOM.render(
-  <ProductsProvider>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </ProductsProvider>,
-  document.getElementById('root')
+configureProductsStore();
+
+const root = ReactDOM.createRoot(document.getElementById("root"));
+
+root.render(
+	// <ProductsProvider>
+	<BrowserRouter>
+		<App />
+	</BrowserRouter>
+	// </ProductsProvider>
 );
