@@ -1,24 +1,28 @@
-import React, { useContext } from 'react';
+import React, { useContext } from "react";
 
-import ProductItem from '../components/Products/ProductItem';
-import { ProductsContext } from '../context/products-context';
-import './Products.css';
+import ProductItem from "../components/Products/ProductItem";
+// import { ProductsContext } from '../context/products-context';
+import { useStore } from "../hooks-store/store";
+import "./Products.css";
 
 const Products = props => {
-  const productList = useContext(ProductsContext).products;
-  return (
-    <ul className="products-list">
-      {productList.map(prod => (
-        <ProductItem
-          key={prod.id}
-          id={prod.id}
-          title={prod.title}
-          description={prod.description}
-          isFav={prod.isFavorite}
-        />
-      ))}
-    </ul>
-  );
+	// const productList = useContext(ProductsContext).products;
+	const state = useStore()[0];
+	return (
+		<ul className="products-list">
+			{/* {productList.map(prod => ( */}
+			{state.products.map(prod => (
+				<ProductItem
+					key={prod.id}
+					id={prod.id}
+					title={prod.title}
+					description={prod.description}
+					isFav={prod.isFavorite}
+				/>
+			))}
+		</ul>
+	);
 };
 
 export default Products;
+
